backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyMySQL==1.1.0
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.0
